hardware/requirements.txt
picamera
RPi.GPIO
schedule
boto3
